requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
fastembed==0.5.1
apache-airflow-providers-weaviate==3.0.3
apache-airflow-providers-snowflake==5.8.1